requirements.txt
aiofile==3.8.8
aiogram[i18n]==3.8.0
asyncpg==0.29.0
cachetools==5.3.3
httpx==0.27.0
litestar[standard]==2.9.1
pyyaml==6.0.1
requests==2.32.3
SQLAlchemy==2.0.31
uvicorn==0.30.1
